clearing some var name to be understandable
